+++ mobile/android/android-components/components/service/pocket/src/main/java/mozilla/components/service/pocket/PocketStoriesConfig.kt
@@ -9,6 +9,7 @@ import mozilla.components.support.base.worker.Frequency
 import java.util.UUID
 import java.util.concurrent.TimeUnit
 
+internal const val DEFAULT_SPONSORED_STORIES_SITE_ID = "1240699"
 internal const val DEFAULT_REFRESH_INTERVAL = 4L
 internal const val DEFAULT_SPONSORED_STORIES_REFRESH_INTERVAL = 4L
 
@@ -26,6 +27,7 @@ internal val DEFAULT_SPONSORED_STORIES_REFRESH_TIMEUNIT = TimeUnit.HOURS
  * @param profile Optional - The profile used for downloading sponsored Pocket stories.
  * @param sponsoredStoriesRefreshFrequency Optional - The interval at which to try and refresh sponsored stories.
  * Defaults to 4 hours.
+ * @param sponsoredStoriesParams Optional - Configuration containing parameters used to get the spoc content.
  */
 class PocketStoriesConfig(
     val client: Client,
@@ -38,7 +40,16 @@ class PocketStoriesConfig(
         DEFAULT_SPONSORED_STORIES_REFRESH_INTERVAL,
         DEFAULT_SPONSORED_STORIES_REFRESH_TIMEUNIT,
     ),
+    val sponsoredStoriesParams: PocketStoriesRequestConfig = PocketStoriesRequestConfig(),
 )
+
+/**
+ * Configuration for sponsored stories request indicating parameters used to get spoc content.
+ *
+ * @property siteId Optional - ID of the site parameter, should be used with care as it changes the
+ * set of sponsored stories fetched from the server.
+ */
+class PocketStoriesRequestConfig(val siteId: String = DEFAULT_SPONSORED_STORIES_SITE_ID)
 
 /**
  * Sponsored stories configuration data.