+++ mobile/android/android-components/components/service/pocket/src/test/java/mozilla/components/service/pocket/PocketStoriesConfigTest.kt
@@ -53,4 +53,11 @@ class PocketStoriesConfigTest {
     fun `GIVEN a Frequency THEN its visibility is internal`() {
         assertClassVisibility(Frequency::class, KVisibility.PUBLIC)
     }
+
+    @Test
+    fun `WHEN instantiating a PocketStoriesConfig THEN sponsoredStoriesParams default value is used`() {
+        val config = PocketStoriesConfig(mock())
+
+        assertEquals(DEFAULT_SPONSORED_STORIES_SITE_ID, config.sponsoredStoriesParams.siteId)
+    }
 }